+++ connect/connect_options.go
@@ -2,19 +2,67 @@ package connect
 
 import (
 	"github.com/yandex-cloud/ydb-go-sdk/v2"
+	"github.com/yandex-cloud/ydb-go-sdk/v2/auth/iam"
+	"context"
+	"fmt"
 	"time"
 )
 
 type ConnectOption func(client *Connection) error
 
-func WithCredentials(credentials ydb.Credentials) ConnectOption {
+func WithAccessTokenCredentials(accessToken string) ConnectOption {
+	return WithCredentials(
+		ydb.NewAuthTokenCredentials(accessToken, "connect.WithAccessTokenCredentials(accessToken)"), // hide access token for logs
+	)
+}
+
+func WithAnonymousCredentials() ConnectOption {
+	return WithCredentials(
+		ydb.NewAnonymousCredentials("connect.WithAnonymousCredentials()"),
+	)
+}
+
+func WithMetadataCredentials(ctx context.Context) ConnectOption {
+	return WithCredentials(
+		iam.InstanceServiceAccount(
+			ydb.WithCredentialsSourceInfo(ctx, "connect.WithMetadataCredentials(ctx)"),
+		),
+	)
+}
+
+func WithServiceAccountKeyFileCredentials(serviceAccountKeyFile string) ConnectOption {
+	return withCredentials(func() (ydb.Credentials, error) {
+		credentials, err := iam.NewClient(
+			iam.WithServiceFile(serviceAccountKeyFile),
+			iam.WithDefaultEndpoint(),
+			iam.WithSystemCertPool(),
+			iam.WithSourceInfo("connect.WithServiceAccountKeyFileCredentials(\""+serviceAccountKeyFile+"\")"),
+		)
+		if err != nil {
+			return nil, fmt.Errorf("configure credentials error: %w", err)
+		}
+		return credentials, nil
+	})
+}
+
+func withCredentials(createCredentials func() (ydb.Credentials, error)) ConnectOption {
 	return func(c *Connection) error {
 		if c.driverConfig == nil {
 			c.driverConfig = &ydb.DriverConfig{}
 		}
+		credentials, err := createCredentials()
+		if err != nil {
+			return err
+		}
 		c.driverConfig.Credentials = credentials
 		return nil
 	}
+}
+
+func WithCredentials(credentials ydb.Credentials) ConnectOption {
+	return withCredentials(func() (ydb.Credentials, error) {
+		return credentials, nil
+	})
 }
 
 func WithDriverConfig(config *ydb.DriverConfig) ConnectOption {