+++ example/auth/access_token_credentials/snippet.go
@@ -1,0 +1,36 @@
+package main
+
+import (
+	"github.com/yandex-cloud/ydb-go-sdk/v2/connect"
+	"context"
+	"flag"
+	"fmt"
+
+	"github.com/yandex-cloud/ydb-go-sdk/v2/example/internal/cli"
+)
+
+type Command struct {
+	accessToken string
+}
+
+func (cmd *Command) Run(ctx context.Context, params cli.Parameters) error {
+	connectCtx, cancel := context.WithTimeout(ctx, params.ConnectTimeout)
+	defer cancel()
+	db, err := connect.New(
+		connectCtx,
+		params.ConnectParams,
+		connect.WithAccessTokenCredentials(cmd.accessToken),
+	)
+	if err != nil {
+		return fmt.Errorf("connect error: %w", err)
+	}
+	defer db.Close()
+
+	// work with db instance
+
+	return nil
+}
+
+func (cmd *Command) ExportFlags(_ context.Context, flagSet *flag.FlagSet) {
+	flagSet.StringVar(&cmd.accessToken, "ydb-access-token", "", "access token for YDB authenticate")
+}