+++ example/auth/metadata_credentials/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/yandex-cloud/ydb-go-sdk/v2/example/internal/cli"
+)
+
+func main() {
+	cli.Run(new(Command))
+}